- Converted all the schema based tests over still missing 100% coverage
