+++ packages/utils/src/index.ts
@@ -30,87 +30,7 @@ import toConstant from './toConstant';
 import toDateString from './toDateString';
 import utcToLocal from './utcToLocal';
 
-import {
-  ArrayFieldTemplateItemType,
-  ArrayFieldTemplateProps,
-  CustomValidator,
-  DateObject,
-  DescriptionFieldProps,
-  ErrorListProps,
-  ErrorSchema,
-  ErrorTransformer,
-  Field,
-  FieldError,
-  FieldErrors,
-  FieldId,
-  FieldPath,
-  FieldProps,
-  FieldTemplateProps,
-  FieldValidation,
-  FormValidation,
-  GenericObjectType,
-  IChangeEvent,
-  IdSchema,
-  ObjectFieldTemplatePropertyType,
-  ObjectFieldTemplateProps,
-  PathSchema,
-  RangeSpecType,
-  Registry,
-  RegistryFieldsType,
-  RegistryWidgetsType,
-  RJSFSchema,
-  RJSFValidationError,
-  SchemaUtilsType,
-  TitleFieldProps,
-  UiSchema,
-  UIOptionsType,
-  UISchemaSubmitButtonOptions,
-  ValidationData,
-  ValidatorType,
-  Widget,
-  WidgetProps,
-} from './types';
-
-export type {
-  ArrayFieldTemplateItemType,
-  ArrayFieldTemplateProps,
-  CustomValidator,
-  DateObject,
-  DescriptionFieldProps,
-  ErrorListProps,
-  ErrorSchema,
-  ErrorTransformer,
-  Field,
-  FieldError,
-  FieldErrors,
-  FieldId,
-  FieldPath,
-  FieldProps,
-  FieldTemplateProps,
-  FieldValidation,
-  FormValidation,
-  GenericObjectType,
-  IChangeEvent,
-  IdSchema,
-  ObjectFieldTemplatePropertyType,
-  ObjectFieldTemplateProps,
-  PathSchema,
-  RangeSpecType,
-  Registry,
-  RegistryFieldsType,
-  RegistryWidgetsType,
-  RJSFSchema,
-  RJSFValidationError,
-  SchemaUtilsType,
-  TitleFieldProps,
-  UiSchema,
-  UIOptionsType,
-  UISchemaSubmitButtonOptions,
-  ValidationData,
-  ValidatorType,
-  Widget,
-  WidgetProps,
-};
+export * from './types';
 
 export * from './constants';
 export * from './schema';
