+++ packages/utils/src/schema/isMultiSelect.ts
@@ -3,7 +3,7 @@ import { RJSFSchema, ValidatorType } from '../types';
 import isSelect from './isSelect';
 
 export default function isMultiSelect<T = any>(
-  validator: ValidatorType, schema: RJSFSchema, rootSchema: RJSFSchema = {}
+  validator: ValidatorType, schema: RJSFSchema, rootSchema?: RJSFSchema
 ) {
   if (!schema.uniqueItems || !schema.items || typeof schema.items === 'boolean') {
     return false;
